다이렉트X 팀프로젝트 총 합본 업로드 합니다
총합본 이메일로도 보냈고 깃헙에도 올려놓습니다. 만점 개꿀!

+++ Engine/TextClass.cpp
@@ -544,7 +544,7 @@ bool TextClass::SetNumOfObjects(int num, ID3D11DeviceContext* deviceContext)
 }
 
 
-bool TextClass::SetSpeed(int cpu, ID3D11DeviceContext* deviceContext)
+bool TextClass::SetSpeed(float cpu, ID3D11DeviceContext* deviceContext)
 {
 	char tempString[16];
 	char cpuString[16];
